Exceptions

+++ src/main/java/com/platform/movies/persistence/MovieEntityRepository.java
@@ -2,6 +2,8 @@ package com.platform.movies.persistence;
 
 import com.platform.movies.domain.dto.MovieDto;
 import com.platform.movies.domain.dto.UpdateMovieDto;
+import com.platform.movies.domain.exception.MovieAlreadyExistsException;
+import com.platform.movies.domain.exception.MovieDontExistsExeption;
 import com.platform.movies.domain.repository.IMovieRepository;
 import com.platform.movies.persistence.entity.MovieEntity;
 import com.platform.movies.persistence.mapper.IMovieMapper;
@@ -26,17 +28,31 @@ public class MovieEntityRepository implements IMovieRepository {
 
     @Override
     public Optional<MovieDto> getMovieById(long id) {
+
+        if (crudMovieEntity.findById(id).isEmpty()) {
+            throw new MovieDontExistsExeption(id);
+        }
+
         return crudMovieEntity.findById(id).map(movieMapper::toDto);
     }
 
     @Override
     public MovieDto save(MovieDto movieDto) {
+        if (crudMovieEntity.findFirstByTitulo(movieDto.title()) != null) {
+            throw new MovieAlreadyExistsException(movieDto.title());
+        }
+
         MovieEntity movieEntity = movieMapper.toEntity(movieDto);
         return movieMapper.toDto(crudMovieEntity.save(movieEntity));
     }
 
     @Override
     public Optional<MovieDto> update(long id, UpdateMovieDto updateMovieDto) {
+
+        if (crudMovieEntity.findById(id).isEmpty()) {
+            throw new MovieDontExistsExeption(id);
+        }
+
         return crudMovieEntity.findById(id).map(entityExistente -> {
             movieMapper.updateEntityFromDto(updateMovieDto, entityExistente);
             return movieMapper.toDto(crudMovieEntity.save(entityExistente));
@@ -45,6 +61,11 @@ public class MovieEntityRepository implements IMovieRepository {
 
     @Override
     public void deleteMovieById(long id) {
+
+        if (crudMovieEntity.findById(id).isEmpty()) {
+            throw new MovieDontExistsExeption(id);
+        }
+
         crudMovieEntity.deleteById(id);
     }
 }
